// src/app/core/model/columnlist.model.ts
export class ColumnListModel {
  name?: string;
  property?: string;
  visible?: boolean;
  searchable?: boolean;
  searchValue?: any;
  isModelProperty?: boolean;
  displayFn: any;
  datatype?: any;
  searchfield?: any;
  isShowinHideList?: any;
  column_width?: any;
  sortOrder?: any;
  dateFiltertype?: any;
}